Revert premature deprecation of `sources` field for `python_awslambda` (#11216)
While the `sources` field results in lots of issues, we did not offer an adequate replacement - there is more boilerplate now.

We likely should still deprecate the `sources` field, but we can do much better. For example, we can teach the `handler` field to allow for a source file, which we then convert into the handler module. Then, we can use dependency inference to avoid the user having to specify `dependencies`.

Status quo, with the deprecation in place:

```python
python_library()

python_awslambda(
  name="lambda",
  handler="path.to.lambda:func",
  dependencies=["./lambda.py"],
)
```

After, with still deprecating, but with the proposed improvements:

```python
python_library()

python_awslambda(
  name="lambda",
  handler="lambda.py:func",
)
```

This new feature shouldn't go into 2.0, so we for now revert the deprecation.

We still keep the bug fix of enforcing that `sources` is <=1 file.