Cancel sessions rather than killing pantsd (#11223)
### Problem

In order to re-enable concurrent runs in #7654, we can no longer kill the server in order to cancel an ongoing run when the client exits (as that would kill all runs rather than just the run associated with the single cancelled client). Instead, we need to gracefully terminate ongoing runs when a `pantsd` client goes away.

### Solution

Using [new facilities](https://github.com/stuhood/nails/commits/9e805a20948a04aea6c082efbe5a9d95b7b71e31) in `nails` `0.8.0`, we detect that a client has `shutdown` the write half of the connection (or otherwise gone away, via heartbeats), and set a cancellation signal on the `Session` for the associated run. When a `Session` has been marked cancelled, the server will cleanly exit the UI loop, and interrupt any `InteractiveProcess` instances.

The client now uses Rust signal handling to intercept Ctrl+C to decide to cancel the run, and will exit the client entirely on the second signal. Because of this, the `ExceptionSink` is now manually installed, and only in locations that are running a `Scheduler` in-process: the `pantsd` server, or a no-`pantsd` local run.

### Result

In general, the server should stay up much longer, because the client no longer kills it in order to handle Ctrl+C. Fixes and unskips some tests related to Ctrl+C handling, and deletes some others. Should cause interactive processes to be consistently interruptible without killing the server.

In another PR we will use this same signal to gracefully cancel any work that was happening in the `Graph`, but since only `InteractiveProcess` or `Console` instances can interact with the UI, that change does not block landing this.

Fixes #7554, fixes #7572, fixes #7547, fixes #7457, fixes #7514.

[ci skip-build-wheels]