Increase Pants' python recursion limit by default, and allow it to be overridden. (#11271)
### Problem

Pants uses a default Python recursion limit that makes some build graph shapes fail to cycle detect: see #11201.

### Solution

Increase the default recursion limit, and allow it to be overridden via an environment variable.

Although increasing the recursion limit is not always the best approach to solving a problem, having the capability to do so is frequently useful as a workaround. We also know that the default value is too low for otherwise reasonable graph shapes.

### Result

Fixes #11201.