Implement native Process cache scoping to fix --test-force (#11291)
The use of `UUIDRequest` to conditionally invalidate `Process` execution was error prone, because it required relevant dependent `@rules` to also be marked `@_uncacheable_rule` (in order to create new `UUIDRequest` objects). This resulted in `--test-force` breaking when those rules were refactored in an attempt to support "per-`pantsd`" caching.

Additionally, the naming/semantics of `ProcessScope` were out of alignment with the meaning of `Session` ("one run of Pants") elsewhere.

Replace `UUIDRequest` with a native implementation of `Process` cache scoping centered around:
```
pub enum ProcessCacheScope {
  // Cached in all locations, regardless of success or failure.
  Always,
  // Cached in all locations, but only if the process exits successfully.
  Successful,
  // Cached only in memory (i.e. memoized in pantsd), but never persistently.
  PerRestart,
  // Never cached anywhere: will run once per Session (i.e. once per run of Pants).
  Never,
}
```

...and add `Process` level tests to validate each of these cases.

Fixes `--test-force` to toggle between the `Never` and `Successful` scopes, and uses `PerRestart` for binary requests.

[ci skip-build-wheels]