Interrupt Graph tasks on Ctrl+C (#11322)
### Problem

As described in #11252, we don't currently cancel work running in the `Graph` when `Scheduler::execute` observes `Session::cancelled` (which is set when a client goes/is-going away due to `Ctrl+C`).

### Solution

With the new interest-based cancellation from #11308, cancelling work on the `Graph` only requires the caller to "go away". So removing the call to `spawn` in `Scheduler::execute` causes `Graph` tasks to be canceled when the loop exits.

### Result

Killing the client of a `pantsd` run with `Ctrl+C` (or in any other way) causes the serverside work to be canceled. Fixes #11252.

[ci skip-build-wheels]