refactor all gRPC usages to use Tonic instead of grpcio (#11307)
### Problem

In testing the remote cache code against Toolchain's remote cache cluster (with both an AWS ALB load balancer and nginx ingress), the Pants client was erroring when receiving HTTP/2 GOAWAY frames [which are allowed by the HTTP/2 standard in ordinary course of shutting down a connection](https://tools.ietf.org/html/rfc7540#section-6.8). The GOAWAY frame should not cause an error when there is no error attached since it is just the server closing the connection, which it is allowed to do. The client should just reconnect.

Pants uses a very old version of the grpcio library (v0.5.x). Thus, this issue could potentially be a bug in grpcio but could have been fixed in a more recent 1.x series release. Long term, we want to switch to Tonic so instead of upgrading grpcio and hoping the issue is fixed, the better use of time is to just port Pants to Tonic.

### Solution

Port all gRPC usage in Pants to the Tonic library and the Prost protobuf code generator. This provides idiomatic Rust bindings for gRPC and protobuf and is directly integrated into tokio and hyper.

### Result

Existing tests continue to pass. Manual testing of remote caching of Toolchain's internal repository continues to work.