Refactor generate_docs.py by extracting out three free functions
ReferenceGenerator was getting pretty crowded. We need to add another thing not really related to the class, so it will be cleaner to be a free function.

Also, now you can follow `main()` to see the logic at a high level. There's less nesting, and side effects are more centralized.

# Rust tests and lints will be skipped. Delete if not intended.
[ci skip-rust]

# Building wheels and fs_util will be skipped. Delete if not intended.
[ci skip-build-wheels]