Add goals to run_info (#11374)
Add a new key "goals" containing the list of goals executed in this pants run, to the `run_info` structure in `RunTracker`, which will be output when the `run_information` method is called. This will allow us to get rid of the slightly-messy code around `_v2_goal_rule_names`, which contains the same information.