Deprecate Subsystems as global singletons (#11388)
### Problem

In order to allow multiple concurrent runs in the same process (#7654), we must remove all usage of global mutable singletons. This includes the "global" and "scoped" `Subsystem` instances (which used to be the only way to consume `Subsystem`s, but which have been replaced by access via the engine).

### Solution

Deprecate `Subsystem.{global_instance,scoped_instance}()`, and implement alternative APIs for:
1. accessing the `RunTracker` from a `StreamingWorkunitHandler` callback
2. registering `StreamingWorkunitHandler` callbacks

The former is accomplished by adding the `RunTracker` to `StreamingWorkunitContext`, and the latter by adding a `@union WorkunitsCallbackFactoryRequest` that can be used to install a `@rule` to construct a `WorkunitsCallbackFactory`.

### Result

In a future version, `Subsystem.{global_instance,scoped_instance}()` will be removed. #7654 gets a little bit closer.

[ci skip-build-wheels]