Remove more obsolete code from RunTracker (#11408)
### Problem

`RunTracker` still has a lot of code paths many years old that have been obsolete for a long time. 

### Solution

This commit consolidates several methods on `RunTracker` that are being called in only one place, adds type annotations to functions that did not previously have them, and removing code that pertains to the old Python-only `WorkUnit` class (unrelated to Rust-level engine workunits). 

### Result

Nothing in this change should be user-visible.