Fix older versions of `tar` not working when creating an `archive` (#11413)
A user reported that the builtin `tar` on macOS Mojave (2.8) was erroring because of the argv including `a`. We instead can explicitly specify the format.

Note that not all formats will still work on older tar versions, e.g. 2.8 does not support the `J` flag to create `.tar.xz`. The user will need to choose a compatible `format` instead like `tar`. We do not eagerly detect this situation because there are so many implementations of `tar` - it's an open question if we should check the failed process, though, so that we can augment the error message with info on this possible gotcha?

[ci skip-rust]
[ci skip-build-wheels]