Add `--remote-cache-eager-fetch` to reduce unreliability (#11396)
Closes #11330. As explained there, lazy fetching is best for performance, but as currently factored, it can result in unrecoverable errors when remote caching fails, e.g. a digest is missing.

By eagerly fetching, we limit the surface area of where remote caching is used; this allows for our error catching logic to always be used, which gracefully falls back to re-running the process locally.

This is likely a temporary solution and this option will be likely deprecated in the future. The more robust solution is to implement https://github.com/pantsbuild/pants/pull/11331.