Remove python workunits (#11416)
### Problem

The Python `workunts.py` module is vestigial at this point. It was designed for the v1 execution model, and is now being used only incidentally to provide a bit of timing data in an overly-complicated way.

### Solution

Modify the several places in the codebase that use `WorkUnit` incidentally to provide timing data or for global pants failure/success values, and then delete the module entirely. Also remove a little bit more vestigial code from `RunTracker`.

### Result

This is a pure refactoring change that should have no user-facing effects.