add histograms for observation metrics to work unit store (#11185)
### Problem

Besides counter metrics, another useful metric is to just observe a value at different points in execution. These observations can be stored as a histogram relatively efficiently.

### Solution

Add a `record_observation` function on `WorkunitStore` that records histograms of observed values. This PR uses the [HDR Histogram library](https://github.com/HdrHistogram/HdrHistogram_rust) which is designed to be fast and efficient. Moreover, the project has libraries for several other languages so we will be able to process and display histograms in languages other than Rust. 

Introduce an observation for the time to first response from REAPI servers (plus a "test observation" for use in an integration test).

The histograms (and not the raw values) are exposed to Python via the `session_get_observation_histograms` FFI function.

### Result

Updated an integration test to generate a test observation and ensure that a histogram appears for it.