Remove --stats-json-file (#11435)
In a recent commit, the RunTracker-specific option to write build stats to a JSON file was deprecated and its functionality moved to a new global option `--stats-json-file`. But in modern pants the ability to write this JSON file isn't particularly useful, so we might as well remove it and the code that supports it.