Add failure logging to the StoreGCService. (#11460)
### Problem

Production `pantsd` logs sometimes indicate that the `StoreGCService` has exited, which triggers a `pantsd` restart. But we do not have enough logging in shutdown to determine the reason for the exit. 

### Solution

We already have test coverage that rules out deterministic failures (and local checks confirm that the store is working as intended in common usage), so add logging to catch the unexpected cases.

[ci skip-rust]
[ci skip-build-wheels]