Remove `RunInfo` class (#11456)
`RunInfo` is not particularly useful anymore; it is basically a complicated way of storing the state of about 10 key-value pairs with information about the current pants run within `RunTracker`. This commit replaces the `RunInfo` code in `RunTracker` with an ordinary Python `dict`, adds a unit test to ensure that `run_information` still yields reasonable results, and deletes the `RunInfo` code entirely.