add "time to first byte" metric for remote CAS reads (#11463)
Add an observation metric to instrument read latency for remote CAS reads. The metric measures the time from the submission of the RPC to receipt of the first byte of the response. This is a better measure of latency since total latency of CAS depends on blob size whereas this just measures the roundtrip time to receive the first byte.