fix use of remote store when in eager cache mode (#11468)
### Problem

As described in https://github.com/pantsbuild/pants/issues/11455, in remote cache mode with eager fetching, the only accesses to the remote CAS should be through the remote cache code; the remainder of the Pants code base should only see the local store. Intrinsics, such as download_file_to_digest, still have access to the remote CAS currently, which means that any issues with the remote CAS will cause an error instead of being a warning if the access to the remote CAS had occurred through the remote cache code. The intended behavior is for all problems with the remote CAS to be warnings when in remote cache mode.

### Solution

In remote cache mode with eager fetching, only expose the local store to most of the Pants code base.

### Result

Added an integration test to test that warnings are generated in remote cache mode.

Fixes https://github.com/pantsbuild/pants/issues/11455