Change logging level for Scrubbed PYTHONPATH from env message + use fstrings (#11480)
### Problem

If PYTHONPATH is defined in the env, this message will show up every time pants start, which is somewhat noisy (and also probably useless)

### Solution

Change log level of that message. Users already expect that Pants is isolating their pythonpath, and (should) know that they need to declare deps.

### Result

Cleaner output in pants runs.