Refactor setting up headers for remote execution (#11500)
This deduplicates converting the BTreeMap of headers into a Tonic interceptor.

We also move the conversion of the oauth_bearer_token_path one level higher in the abstraction chain. In a followup, we will move this outside of Rust entirely to be handled by Python. Then, Python will simply pass a dictionary of headers to Rust, and Rust will be agnostic to how it was created. This will facilitate adding a new plugin hook for adding authentication for remote execution/caching.