Add `--remote-store-headers` and fix `--remote-execution-headers` to not impact remote caching (#11501)
Previously, we only had `--remote-execution-headers`. We weren't applying those to the store setup, and we were incorrectly applying them to the remote cache setup.

We considered instead consolidating to only have `--remote-headers`, which applies to both contexts. However, prior art from Bazel shows that it's useful to have service-specific options and we avoid a deprecation warning this way. We could add both `--remote-headers` and `--remote-store-headers`, but that's not done here for simplicity. We can add `--remote-headers` in the future, if necessary.

This PR also refactors to convert the `--remote-oauth-bearer-token-path` into the relevant header in Python, before crossing the FFI boundary. This simplifies our Rust code so that it simply gets a dictionary of headers. All header injection now happens in Python, which will facilitate adding a plugin hook to dynamically set these headers.