Stop pinning `zipp` by default in `[pytest].pytest_plugins` (#11512)
Pip's new resolver now errors when a constraints.txt value != an explicit requirement, such as one provided via `[pytest].pytest_plugins`. This means that if a user has `zipp` in their constraints.txt, they must either pin to the old `2.1.0` or they must override the default for `[pytest].pytest_plugins`. Both are very non-obvious what to do.

We had pinned zipp as a convenience because it was unstable for some time, but it seems to have become more stable. It's been working robustly to use zipp 3.* in Toolchain. Now, this is likely to be much more of a gotcha and nuisance than a value-add to users. So, we stop pinning.

[ci skip-rust]
[ci skip-build-wheels]