Use named field struct syntax for Digest (#11492)
### Problem

We might want to add an additional field on `Digest` in order to support correctly handling `Digest`s created from a file (a Python `FileDigest`) vs `Digest`s created from a directory (a Python `Digest` or `Snapshot`). Since this would involve changing the same code in a bunch of different places in the codebase where we create a `Digest` literal, preemptively change the syntax to named-field-struct-style so this only needs to happen once.

### Solution

Change the syntax and add a `::new()` associated function to avoid having to reference the literal struct syntax all over the codebase.