Put test stdout/stderr onto EnrichedTestResult workunit (#11528)
### Problem

We created the `EnrichedTestResult` wrapper type - which is returned from an uncacheable rule - in order to ensure that pants runs would always have a workunit with test output metadata on it, even if the test had been cached. However, the stdout and stderr digests from the test were still being put on the underlying process execution workunit, which doesn't get created when the process result is cached.

### Solution

In order to do this correctly, the engine's treatment of `Digest`s had to be modified. Some `Digest`s, like the stdout/stderr output ones, represent just a single conceptual file and correspond to a Python `FileDigest`; and others represent a file-system subtree and correspond to a Python `Snapshot`. The code that handles workunit artifacts in the engine needs to keep these two interpretations of the Rust `hashing::Digest` type straight to avoid runtime errors converting and deconverting from Python types, which was done by introducing the `ArtifactResult` wrapper type. 

As well, the signature of `EngineAwareReturnType.artifacts` was modified to be a dictionary of `Union[Snapshot, FileDigest]` rather than just `Snapshot`. Note that the custom code to put the stdout/stderr digests on the process workunit was previously violating the type contract, by placing `FileDigest` Python Values into a dictionary that was supposed to only contain `Snapshot`s - but because this was done in engine FFI code, there was no opportunity for MyPy to verify this.