Support --venv mode for internal PEXes.
Add rules to generate a VenvPex that is a normal PEX file that operates
in --venv mode. To support native venv execution speeds, the VenvPex uses
a wrapper script that runs directly from a seeded venv when available and
re-seeding when not.

Leverage these rules to speed up and simplify all internal PEX usages.

Fixes #11531

# Rust tests and lints will be skipped. Delete if not intended.
[ci skip-rust]

# Building wheels and fs_util will be skipped. Delete if not intended.
[ci skip-build-wheels]