Simplify native code bootstrapping.
Previously we cached the native_engine.so and its metadata
under ~/.cache/pants/bin/native-engine, and we performed the
hash checks against the cached copies, requiring them to exist
even if the actual resource already exists at the right hash.

This is unintuitive, as you'd expect that deleting a cache shouldn't
matter if the underlying resource is up to date in its final location.
For example, complicates CI, as it requires that cache directory to be
conserved across runs to avoid a full rebuild, even if the
resource is conserved.

Furthermore, the cache itself is unnecessary. Cargo will rebuild
incrementally, and without Rust code changes the invocation only
takes ~250 ms. So the only value of the cache is to avoid an
incremental rebuild when restoring to an early version of the Rust
code that was already built and cached. That alone does not seem
worth the complexity.

This change eliminates that caching and simplifies bootstrapping to
merely ensure that the final resource is up to date.

[ci skip-build-wheels]