Fix missing counters by adding more workunits (#11555)
As found in https://github.com/pantsbuild/pants/issues/11548, the remote cache write counters were missing because they referred to a workunit that was already complete. This works around that by adding a new workunit in the async block.

We also add a new workunit for local cache reads. Technically, we don't need this, but it will be necessary if/when we land the proposed fix in https://github.com/pantsbuild/pants/issues/11548 because we won't have a way of passing the parent workunit to the `ComandRunner.run()` method, given its trait signature that we can't change.