Expose all counter names to streaming workunit handlers (#11573)
Currently, counters are only reported if their value is >0. However, it can be useful for consumers to know when a counter is set to 0, e.g. having positive confirmation that there were 0 remote caching errors.

We still keep individual workunits as only recording what counters were actually used; this keeps that code simpler and avoids unnecessary allocations.

Instead, we add a new property to `RunTracker` to access a list of all the counter names. Consumers can use this to then backtrack what is missing, so infer the count is 0. The `RunTracker` can be accessed via `StreamingWorkunitContext`.