Renames brfs tests to have a `brfs_` prefix.
This makes it easy to skip them using the `--skip` flag.

BRFS is not currently used, and we may want to skip its
tests in MacOS CI, since getting an environment with
osxfuse working is tricky. The linux tests will suffice
to avoid bitrot until such times as we rejuvenate BRFS
or get rid of it.

However we don't want to use the #[ignore] facility for
this since we don't want to skip these tests by default
(e.g., we don't want to skip them on CI linux runs, or
local MacOS runs). We want to exlude tests only very explicitly.

[ci skip-build-wheels]