Ensure that ancestor files at the buildroot are found.
Previously we would find ancestor files correctly all the
way to the source root, and even above, but would stop short
at the buildroot.

However there are real-world cases in which the buildroot is
the source root, and can contain, e.g., conftest.py files which
we fail to discover.

This change fixes this oversight.

[ci skip-rust]

[ci skip-build-wheels]