Remove WRAPPER_REQUIREMENTS and WRAPPER_SRCPATH (#11645)
This commit removes the code in the pants-repo-specific `pants` script that modifies the pants Python environment based on the `WRAPPER_REQUIREMENTS` and `WRAPPER_SRCPATH` variables. This functionality doesn't seem to have been used in many years, and removing it makes the `pants` script easier to understand, which is useful from the perspective of making changes to pants' initialization process (e.g. eventually running pants code via an embedded Python interpreter).