Fix streaming workunit handler plugins to log to console during run (#11674)
While the final `poll_workunits()` call would write to the console when `finished=True`, any runs before would write to `.pants.d/pants.log` because they were running on a background thread and using the default destination of the logfile. This means that logs plugin authors expected to show to the user would not render.

This is fixed by using our Rust functions to set the correct thread destination for the current thread. We expose an FFI function to get the parent thread's destination, then use it to set the destination once the new thread has started.

Example of this working by adding a log to `--stats-log`:

```
❯ ./pants --stats-log --no-process-execution-use-local-cache lint src/python/pants:
15:51:34.77 [INFO] initializing scheduler...
15:51:35.92 [INFO] Starting: Resolving plugins: hdrhistogram
15:51:37.53 [INFO] Completed: Resolving plugins: hdrhistogram
15:51:37.98 [INFO] scheduler initialized.
15:51:38.01 [INFO] HERE!
15:51:39.01 [INFO] HERE!
15:51:40.01 [INFO] HERE!
15:51:40.06 [INFO] Completed: Building docformatter.pex with 1 requirement: docformatter>=1.3.1,<1.4
```

Unblocks https://github.com/pantsbuild/pants/issues/11433.