Fix time zones github actions (#11682)
The Github Actions docs say that the time in the `on: schedule:` specifier is in UTC, but one of them was written at a time which is late at night in UTC and therefore in the afternoon in Pacific Time. This commit makes the scheduled times for both the daily extended testing and `cargo audit` jobs be at times in UTC that correspond to times late at night in Pacific Time, and clarifies the comments around this.