Refactor `StreamingWorkunitHandler` to be a class-based context manager (#11685)
Before, it was possible to call `StreamingWorkunitHandler.session()` multiple times, meaning we'd be starting and stopping the same threads during the same Pants run. While this wasn't done in practice beyond a test, it complicates future changes and is an inaccurate representation of what we expect.

Even though it is still possible to do this with a class-based context manager, the class better expresses the intention than a reusable method. Further, this change will result in a RuntimeException if the user does use the context manager twice.

[ci skip-rust]
[ci skip-build-wheels]