Send anonymous telemetry.
Uses the `humbug` library to send anonymized telemetry data
to bugout.

Repos must opt in to telemetry. We log a warning if a repo has
not opted in or out.

See the code and also https://www.pantsbuild.org/v2.4/docs/anonymous-telemetry
for more info on the data we send and how we preserve anonymity
and prevent leakage of proprietary information.

[ci skip-rust]

[ci skip-build-wheels]