Improve error message for unmatched globs (#11699)
Closes https://github.com/pantsbuild/pants/issues/11629. Example:

```
❯ ./pants --no-print-stacktrace lint src/fake.txt --no-pantsd
16:03:43.20 [ERROR] Exception caught: (pants.engine.internals.scheduler.ExecutionError)
...
Exception message: 1 Exception encountered:

  Exception: Unmatched glob from file arguments: "src/fake.txt"

Does the file exist? If so, check if the file is in your `.gitignore` or the global `pants_ignore` option, which may result in Pants not being able to see the file even though it exists on disk. Refer to https://www.pantsbuild.org/v2.4/docs/troubleshooting#pants-cannot-find-a-file-in-your-project.
```